mail and cart

+++ market/database/migrations/2024_12_05_193337_add_category_id_to_products_table.php
@@ -9,14 +9,15 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up()
+    public function up(): void
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->unsignedBigInteger('category_id')->after('price');
-            $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
+            if (!Schema::hasColumn('products', 'category_id')) {
+                $table->unsignedBigInteger('category_id')->after('price');
+                $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
+            }
         });
     }
-    
 
     /**
      * Reverse the migrations.
@@ -24,7 +25,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('products', function (Blueprint $table) {
-            //
+            if (Schema::hasColumn('products', 'category_id')) {
+                $table->dropForeign(['category_id']);
+                $table->dropColumn('category_id');
+            }
         });
     }
 };
